requirements.txt
tweepy==4.14.0
praw==7.7.1
redis==5.0.1
python-dotenv==1.0.0
anthropic==0.25.0
schedule==1.2.1
solana==0.36.6
solders==0.26.0
anchorpy==0.21.0
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.0
